move aux functions to md utils, improve user output
